requirements.txt
Flask==3.0.2
flask-cors==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
Flask-Migrate==4.0.5 